requirements.txt
fastapi
uvicorn
websockets
assemblyai
edge-tts
python-dotenv
aiohttp
openai
requests
supabase
